[nomaster] removes duplication in inferImplicitValue
Shame-driven development at its best.
